Do some refactoring on Not found exceptions and idmismatch exceptions

+++ src/main/java/wolox/training/controllers/BookController.java
@@ -4,6 +4,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,8 +17,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import wolox.training.exceptions.EntityNotFoundException;
 import wolox.training.exceptions.IdMismatchException;
-import wolox.training.exceptions.BookNotFoundException;
 import wolox.training.models.Book;
 import wolox.training.repositories.BookRepository;
 
@@ -44,14 +45,16 @@ public class BookController {
     /**
      * Find one book by id.
      *
-     * @throws BookNotFoundException
+     * @throws EntityNotFoundException
      * @param id the id
-     * @return the book
+     * @return the {@link Book}
      */
     @GetMapping("/{id}")
     public Book findOne(@PathVariable Long id) {
-        return bookRepository.findById(id)
-            .orElseThrow(BookNotFoundException::new);
+        Optional<Book> book = bookRepository.findById(id);
+        if (!book.isPresent())
+            throw new EntityNotFoundException(Book.class);
+        return book.get();
     }
 
     /**
@@ -69,14 +72,15 @@ public class BookController {
     /**
      * Delete an existing book.
      *
-     * @throws BookNotFoundException
+     * @throws EntityNotFoundException
      * @param id the path variable id of the book to find and delete
      */
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
-        bookRepository.findById(id)
-            .orElseThrow(BookNotFoundException::new);
+        Optional<Book> book = bookRepository.findById(id);
+        if (!book.isPresent())
+            throw new EntityNotFoundException(Book.class);
         bookRepository.deleteById(id);
     }
 
@@ -84,7 +88,7 @@ public class BookController {
      * Updates an existing book.
      *
      * @throws IdMismatchException if pathVariable id does not match body id.
-     * @throws BookNotFoundException
+     * @throws EntityNotFoundException
      * @param book the request updated book to save
      * @param id   the path variable for the book id
      * @return the saved updated book
@@ -100,8 +104,9 @@ public class BookController {
         if (book.getId() != id) {
             throw new IdMismatchException("book");
         }
-        bookRepository.findById(id)
-            .orElseThrow(BookNotFoundException::new);
+        Optional<Book> dbBook = bookRepository.findById(id);
+        if (!dbBook.isPresent())
+            throw new EntityNotFoundException(Book.class);
         bookRepository.save(book);
     }
 }
